Reorganize the API using Blueprints and modularization
