+++ AbbyBot-API/app/utils/discord.py
@@ -1,0 +1,30 @@
+import requests
+import os
+from .db import get_server_count
+
+DISCORD_API_BASE_URL = "https://discord.com/api/v10"
+TOKEN = os.getenv('DISCORD_TOKEN')
+
+def get_bot_info_from_discord():
+    url = f"{DISCORD_API_BASE_URL}/users/@me"
+    headers = {
+        "Authorization": f"Bot {TOKEN}"
+    }
+
+    response = requests.get(url, headers=headers)
+
+    if response.status_code == 200:
+        bot_info = response.json()
+        bot_data = {
+            "bot_id": bot_info["id"],
+            "bot_name": bot_info["username"],
+            "discriminator": bot_info["discriminator"],
+            "avatar_url": f"https://cdn.discordapp.com/avatars/{bot_info['id']}/{bot_info['avatar']}.png",
+            "banner_url": f"https://example.com/bot/banner/{bot_info['id']}.png",  
+            "server_count": get_server_count(),  
+            "version": os.getenv('BOT_VERSION')
+        }
+        return bot_data
+    else:
+        print(f"Error fetching bot info: {response.status_code} {response.text}")
+        return None
